refactor: moved UkrainianTexTemplate to manim_config.py
